Part 3: submit Cardano tx with Pyth Lazer RWA price verification
Fetch latest XAU/USD from Pyth Lazer, resolve on-chain Pyth State,
and submit zero-withdrawal transaction on PreProd to verify the
signed price update on-chain.

Verified tx: 8fca5fdd8e831c5e1ee0d8417cfedae480b6d2a7374647ded100b440fca49e43
