+++ lazer/cardano/lazer-rwa/package.json
@@ -4,7 +4,7 @@
   "type": "module",
   "scripts": {
     "fetch-prices": "npx tsx src/fetch_prices_lazer_rwa.ts",
-    "submit-tx": "npx tsx src/submit_tx.ts"
+    "submit-tx": "npx tsx src/submit_tx_lazer_rwa.ts"
   },
   "dependencies": {
     "@evolution-sdk/evolution": "^0.3.30",
